requirements.txt
Django==2.2.16
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.1
selenium-wire==2.1.1
gunicorn==20.0.4
